Initial Prototype
This version should be able to do the following things:

1) Handle Hardware PWM for the Raspberry Pi
2) Handle I2C for the Raspberry Pi

This is heavily based on SwiftyGPIO, but with more type safety when dealing with the underlying registers, instead of heavily relying on constants.

This is enough for the PCA9685 library and the RPiLight project to start flushing out the bugs.