Fix Write Access to Pins
By making the BoardGPIOPin/BoardGPIOPinSet protocols as a class, the properties are available to be tweaked despite the subscripts being get-only. Otherwise, things like:

gpio[12].value = true

Don’t really work, and Swift does some goofy stuff where it attempts to generate code that gets the BoardGPIOPin, and then set it back after the modification. This is perfect for value types. But these protocols aren’t meant for value types, so we don’t really want that behavior.

While we are at it, make sure the core protocols are all class-only for now. Since this is about access to a single resource, reference semantics are correct.