Remove Some Optionals
Makes a handful of changes with the goal of cutting down how many optionals that a user needs to go through in order to interact with a board.

- Split the “Board” protocol up into specific capabilities. This makes it possible to compose a specific board’s protocol from other protocols. This means if a board doesn’t support a feature, it’s not exposed to the developer. Period. But it still uses protocols to enforce consistent behaviors between boards with the same capabilities.
- Offer up a “main bus” property for I2C capabilities. So instead of having to do something like: board.i2cBus[0]?.doAThing(), you can now do: board.i2cMainBus.doAThing(). Just as many characters, but less need for error checking, and less need to know which bus is the “default” one for a board. For example, Rock64 uses bus 1 for system devices, and bus 0 is available for developer use. On the Pi, bus 0 is for system devices, and bus 1 is for the developer. Ugh.
- Some renaming of types to make things a little clearer to developers trying to make sense of it.

This also paints a path for doing similar in other areas. One example area that needs to be cleaned up for Rock64 support in the future is the GPIO protocols. Right now they assume what type to use for the set of pins. It’d be nice if this was more flexible, or the ability to use a set of pins could be omitted from a board.