Fix up support for basic/raw I2C
Up to this point, SingleBoard only supported SMBus commands over I2C. With these changes, it should support raw I2C a bit better.

It also includes an early implementation of encode/decode of types. This makes it possible to write/read structs/classes directly with a BoardI2CEndpoint, by having the struct encode/decode to a [UInt8] array. It’s not quite as nice as the RawRepresentable support for pure bitfields/enums and other simple register-style layouts, but is required when dealing with more complex compact buffers that are common in I2C.