Clean up I2C API: Replace Arrays with Data
Since we have a window to break the I2C API a little, this chance removes the use of arrays for I2C and instead uses Data objects.

It also renames “from:/to:” to “command:” so things are clearer with the new I2C API to tell the difference between the SMBus and the generic I2C parts of the interface.